pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cccc-pair"
version = "0.4.0a0"
description = "Global multi-agent delivery kernel with working groups, scopes, and an append-only collaboration ledger"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.9"
authors = [
  { name = "ChesterRa", email = "ra@ike.ba" }
]
license = { text = "Apache-2.0" }
keywords = ["orchestrator","ai","rfd","pair","collaboration"]
dependencies = [
  "PyYAML>=6.0",
  "fastapi>=0.110",
  "pydantic>=2.0",
  "uvicorn[standard]>=0.27",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Environment :: Console",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS",
  "Topic :: Software Development :: Build Tools",
]

[project.urls]
Homepage = "https://github.com/ChesterRa/cccc"
Repository = "https://github.com/ChesterRa/cccc"
Issues = "https://github.com/ChesterRa/cccc/issues"

[project.scripts]
cccc = "cccc.cli:main"
ccccd = "cccc.daemon_main:main"

[tool.setuptools]
include-package-data = true
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"cccc.ports.web" = ["dist/**"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["cccc*"]
